+++ Runtime/General/PlayerControlledVelocity.cs
@@ -2,27 +2,28 @@
 using System.Collections.Generic;
 using UnityEngine;
 
-
 namespace Botaemic.Core
 {
-    public class SetConstantVelocity : MonoBehaviour
+    public class PlayerControlledVelocity : MonoBehaviour
     {
         [SerializeField] private Vector3 _velocity = Vector3.zero;
+        [SerializeField] private KeyCode _keyPositive = KeyCode.D;
+        [SerializeField] private KeyCode _keyNegative = KeyCode.A;
 
         private Rigidbody _rb;
 
         #region Unity Functions
-        private void Awake()
+        void FixedUpdate()
         {
-            _rb = GetComponent<Rigidbody>();
-            if (_rb == null) { LogWarning("No rigidbody detected"); }
-        }
+            if (Input.GetKey(_keyPositive))
+                GetComponent<Rigidbody>().velocity += _velocity;
 
-        private void Start()
-        {
-            _rb.velocity = _velocity;
+            if (Input.GetKey(_keyNegative))
+                GetComponent<Rigidbody>().velocity -= _velocity;
         }
         #endregion
+
+
 
         #region Logging
         private void Log(string text)
@@ -32,7 +33,7 @@ namespace Botaemic.Core
 
         private void LogWarning(string text)
         {
-            DebugUtility.Log("WARNING! - " + transform.name + " => " + text);
+            DebugUtility.Log("WARNING! " + text);
         }
         #endregion
     }